Move evaluation checks to shared module
