+++ analyze_website.py
@@ -3,50 +3,17 @@ from bs4 import BeautifulSoup
 import json
 import sys
 
+from ai_readiness.evaluations import (
+    check_semantic_html,
+    check_schema_markup,
+    check_headings_structure,
+    check_alt_text,
+    check_lists_and_tables,
+    check_language_attribute,
+    check_transcripts_captions,
+)
+
 # ========== EVALUATION FUNCTIONS ==========
-
-def check_semantic_html(soup):
-    tags = ["header", "nav", "main", "article", "section", "footer"]
-    found = any(soup.find(tag) for tag in tags)
-    return found, "Semantic HTML tags found." if found else "Add semantic HTML5 tags for structure."
-
-def check_schema_markup(soup):
-    # Simple check: look for JSON-LD scripts containing 'schema.org'
-    schema = soup.find_all("script", type="application/ld+json")
-    found = any("schema.org" in script.text for script in schema)
-    return found, "Schema.org markup present." if found else "Add Schema.org structured data."
-
-def check_headings_structure(soup):
-    headings = soup.find_all(["h1", "h2", "h3", "h4", "h5", "h6"])
-    found = bool(headings)
-    return found, f"{len(headings)} heading tags found." if found else "Add descriptive headings (h1-h6)."
-
-def check_alt_text(soup):
-    images = soup.find_all("img")
-    images_missing_alt = [img for img in images if not img.has_attr('alt') or not img['alt'].strip()]
-    found = len(images_missing_alt) == 0 and len(images) > 0
-    return found, "All images have alt text." if found else f"{len(images_missing_alt)} images missing alt text."
-
-def check_lists_and_tables(soup):
-    lists = soup.find_all(['ul', 'ol'])
-    tables = soup.find_all("table")
-    found = bool(lists or tables)
-    return found, f"{len(lists)} lists, {len(tables)} tables found." if found else "Consider using lists and tables for structured content."
-
-def check_language_attribute(soup):
-    html_tag = soup.find("html")
-    found = html_tag and html_tag.has_attr('lang')
-    return found, "lang attribute present on html tag." if found else "Add lang attribute for language declaration."
-
-def check_transcripts_captions(soup):
-    # For MVP: flag if there are audio/video tags without descriptions/captions
-    videos = soup.find_all("video")
-    audios = soup.find_all("audio")
-    found = True  # assume pass if no multimedia
-    if videos or audios:
-        found = all((v.get('aria-label') or v.get('title') or v.find('track', attrs={'kind': 'captions'})) for v in videos)
-        found = found and all((a.get('aria-label') or a.get('title')) for a in audios)
-    return found, "Multimedia elements have ARIA labels/titles/captions." if found else "Provide captions/transcripts for videos/audio."
 
 def analyze_website(url):
     try:
